compiling project without errors
