+++ rssfeeder-api/pom.xml
@@ -5,11 +5,12 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>org.araa</groupId>
-        <artifactId>RSSFeeder</artifactId>
+        <artifactId>rssfeeder-parent</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
 
     <artifactId>rssfeeder-api</artifactId>
+    <version>${project.parent.version}</version>
 
     <properties>
         <maven.compiler.source>17</maven.compiler.source>
@@ -20,30 +21,44 @@
     <dependencies>
         <dependency>
             <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter</artifactId>
+            <exclusions>
+                <exclusion>
+                    <groupId>org.springframework.boot</groupId>
+                    <artifactId>spring-boot-starter-logging</artifactId>
+                </exclusion>
+            </exclusions>
+        </dependency>
+
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web</artifactId>
         </dependency>
 
         <dependency>
             <groupId>org.araa</groupId>
             <artifactId>rssfeeder-utitly</artifactId>
-            <version>${project.version}</version>
         </dependency>
         <dependency>
             <groupId>org.araa</groupId>
             <artifactId>rssfeeder-domain</artifactId>
-            <version>${project.version}</version>
         </dependency>
         <dependency>
             <groupId>org.araa</groupId>
             <artifactId>rssfeeder-service</artifactId>
-            <version>${project.version}</version>
         </dependency>
         <dependency>
-            <groupId>org.araa</groupId>
-            <artifactId>rssfeeder-security</artifactId>
-            <version>${project.version}</version>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
         </dependency>
-
     </dependencies>
 
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.springframework.boot</groupId>
+                <artifactId>spring-boot-maven-plugin</artifactId>
+            </plugin>
+        </plugins>
+    </build>
 </project>
