+++ rssfeeder-api/src/main/java/org/araa/Feeder.java
