+++ rssfeeder-api/src/main/java/org/araa/dto/UserRegistrationResponseDTO.java
@@ -2,7 +2,6 @@ package org.araa.dto;
 
 import lombok.Builder;
 import lombok.Data;
-import org.araa.domain.User;
 
 
 @Builder
@@ -11,10 +10,4 @@ public class UserRegistrationResponseDTO {
     private String username;
     private String name;
     private String email;
-
-    public UserRegistrationResponseDTO( User user ) {
-        this.username = user.getUsername();
-        this.name = user.getName();
-        this.email = user.getEmail();
-    }
 }
