+++ rssfeeder-persistance/src/main/java/config/RedisCacheConfig.java
@@ -1,4 +1,4 @@
-package org.araa.utility.config;
+package config;
 
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.context.annotation.Bean;
