+++ rssfeeder-security/src/main/java/org/araa/config/JWTAuthenticationFilter.java
@@ -7,8 +7,6 @@ import jakarta.servlet.http.HttpServletResponse;
 import lombok.NonNull;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.araa.application.security.JWTGenerator;
-import org.araa.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -16,6 +14,8 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
 import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
+import org.araa.services.CustomUserDetailsService;
+import org.araa.utility.JWTGenerator;
 
 import java.io.IOException;
 
@@ -26,7 +26,7 @@ public class JWTAuthenticationFilter extends OncePerRequestFilter {
     @Autowired
     private JWTGenerator jwtGenerator;
     @Autowired
-    private UserService userService;
+    private CustomUserDetailsService customUserDetailsService;
 
 
     @Override
@@ -37,7 +37,7 @@ public class JWTAuthenticationFilter extends OncePerRequestFilter {
 
         if ( StringUtils.hasText( token ) && jwtGenerator.validateToken( token ) ) {
             String username = jwtGenerator.getUsernameFromJWT( token );
-            UserDetails userDetails = userService.loadUserByUsername( username );
+            UserDetails userDetails = customUserDetailsService.loadUserByUsername( username );
             UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken( userDetails,
                     null,
                     userDetails.getAuthorities() );
