+++ rssfeeder-service/pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>org.araa</groupId>
-        <artifactId>RSSFeeder</artifactId>
+        <artifactId>rssfeeder-parent</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
 
@@ -19,29 +19,42 @@
 
     <dependencies>
         <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-web</artifactId>
-        </dependency>
-
-        <dependency>
             <groupId>org.araa</groupId>
             <artifactId>rssfeeder-utitly</artifactId>
-            <version>${project.version}</version>
         </dependency>
         <dependency>
             <groupId>org.araa</groupId>
             <artifactId>rssfeeder-domain</artifactId>
-            <version>${project.version}</version>
         </dependency>
         <dependency>
             <groupId>org.araa</groupId>
             <artifactId>rssfeeder-persistance</artifactId>
-            <version>${project.version}</version>
         </dependency>
         <dependency>
             <groupId>org.araa</groupId>
             <artifactId>rssfeeder-security</artifactId>
-            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.junit.jupiter</groupId>
+            <artifactId>junit-jupiter-api</artifactId>
+            <version>5.10.2</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-core</artifactId>
+            <version>4.2.0</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-junit-jupiter</artifactId>
+            <version>5.7.0</version>
+            <scope>test</scope>
         </dependency>
     </dependencies>
