+++ rssfeeder-service/src/main/java/org/araa/services/UserService.java
@@ -13,12 +13,9 @@ import org.araa.domain.User;
 import org.araa.repositories.UserRepository;
 import org.springframework.data.redis.cache.RedisCacheManager;
 import org.springframework.scheduling.annotation.Async;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.security.crypto.password.PasswordEncoder;
 
 import java.util.Date;
 import java.util.List;
@@ -28,7 +25,7 @@ import java.util.concurrent.CompletableFuture;
 
 @AllArgsConstructor
 @Service
-public class UserService implements UserDetailsService {
+public class UserService {
 
     private static final Logger logger = LogManager.getLogger( UserService.class );
     private final UserRepository userRepository;
@@ -55,14 +52,6 @@ public class UserService implements UserDetailsService {
         return user;
     }
 
-    @Override
-    public UserDetails loadUserByUsername( String username ) throws UsernameNotFoundException {
-        User user = userRepository.findByUsername( username )
-                .orElseThrow( () -> new UsernameNotFoundException( username + " not found" ) );
-
-        return new org.springframework.security.core.userdetails.
-                User( user.getUsername(), user.getPassword(), mapRolesToAuthorities( user.getRoles() ) );
-    }
 
 
     @Async
@@ -86,11 +75,6 @@ public class UserService implements UserDetailsService {
         Objects.requireNonNull( redisCacheManager.getCache( "entries" ) ).clear();
     }
 
-    private List<SimpleGrantedAuthority> mapRolesToAuthorities( List<Role> roles ) {
-        return roles.stream()
-                .map( role -> new SimpleGrantedAuthority( "ROLE_" + role.getType() ) )
-                .toList();
-    }
 
 
     public Set<RSS> getUserSubscriptions( User user ) {
