--- src/test/java/org/araa/controllers/RSSControllerTest.java
@@ -1,28 +1,0 @@
-package org.araa.controllers;
-
-import org.araa.services.RSSService;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-@ExtendWith( MockitoExtension.class)
-class RSSControllerTest {
-
-    @Mock
-    private RSSService rssService;
-
-    @InjectMocks
-    private RSSController rssController;
-
-    @Test
-    void registerRSS() {
-//        RSSDto rssDto = rssService.save( "https://www.newswire.lk/feed" );
-
-//        CompletableFuture<ResponseEntity<RSSDto>> response = rssController.save("https://www.newswire.lk/feed");
-//
-//        assertNotNull( response.join().getBody() );
-//        assertEquals( rssDto, response.join().getBody() );
-    }
-}
